+++ spec/rails_app/app/components/grid/associations.rb
@@ -2,7 +2,7 @@ class Grid::Associations < Netzke::Grid::Base
   def configure(c)
     super
     c.model = 'Book'
-    c.edit_inline = true
+    c.editing = :inline
     c.columns = [:title, :author__first_name, :author__prize_count]
   end
 end