+++ spec/rails_app/app/components/grid/crud_inline.rb
@@ -1,6 +1,6 @@
 class Grid::CrudInline < Grid::Crud
   def configure(c)
     super
-    c.edit_inline = true
+    c.editing = :inline
   end
 end