+++ spec/rails_app/app/components/grid/default_values_inline.rb
@@ -1,6 +1,6 @@
 class Grid::DefaultValuesInline < Grid::DefaultValues
   def configure(c)
     super
-    c.edit_inline = true
+    c.editing = :inline
   end
 end