+++ spec/rails_app/app/components/grid/inline_editing.rb
@@ -2,6 +2,6 @@ class Grid::InlineEditing < Netzke::Grid::Base
   def configure(c)
     super
     c.model = 'Author'
-    c.edit_inline = true
+    c.editing = :inline
   end
 end