+++ spec/rails_app/app/components/grid/multiline_edit.rb
@@ -2,6 +2,6 @@ class Grid::MultilineEdit < Netzke::Grid::Base
   def configure(c)
     super
     c.model = "Book"
-    c.edit_inline = true
+    c.editing = :inline
   end
 end