+++ spec/rails_app/app/components/grid/non_paging.rb
@@ -1,0 +1,11 @@
+class Grid::NonPaging < Netzke::Grid::Base
+  def attributes
+    [:title]
+  end
+
+  def configure(c)
+    c.model = 'Book'
+    c.paging = :none
+    super
+  end
+end