+++ spec/rails_app/app/components/grid/non_paging_inline.rb
@@ -1,0 +1,12 @@
+class Grid::NonPagingInline < Netzke::Grid::Base
+  def attributes
+    [:title]
+  end
+
+  def configure(c)
+    c.model = 'Book'
+    c.paging = :none
+    c.editing = :inline
+    super
+  end
+end