+++ spec/rails_app/app/components/grid/paging.rb
@@ -1,7 +1,7 @@
 class Grid::Paging < Netzke::Grid::Base
   def configure(c)
     c.model = 'Book'
-    c.paging = true
+    c.paging = :pagination
     c.store_config = { page_size: 2 }
     c.columns = [:title]
     super